Fixed a long-standing race condition, that was very rare but still happened occasionally. The race was related to the fact that, in preemptive contexts, expressions such as  "cctx[cpu_core_id].current_thread" can fail.
This happens if a thread is relocated to a different core,
between the time cpu_core_id is read and the rest of the expression
is computed.

To do this, the implementation of VM interrupts had to receive a major
rehaul. Now, we are using signal masking of SIGUSR1 to disable interrupts.
Unfortunately, the changes increase the cost of system time by about
25%. :-()

A number of changes had to be made to scheduler and concurrency control,
in order to accomodate the new code with less cost. Overall the new
code is simpler (!) but slower. This may actually be a feature ...